Poll Request get Requests Id

+++ src/main/java/org/cnc/cncbot/dto/serverinfos/ServerInfoResponse.java
@@ -128,7 +128,7 @@ public class ServerInfoResponse {
 
 	private Integer npcufrp;
 
-	private Integer npcurp;
+	private Double npcurp;
 
 	private Double npcurpb;
